More bits and pieces, refactoring websockets
